requirements.txt
anyio==3.5.0
asgiref==3.5.1 
click==8.1.3   
colorama==0.4.4
fastapi
h11==0.13.0    
idna==3.3      
imbalanced-learn==0.9.0
imblearn==0.0
joblib==1.1.0
numpy==1.22.3
pandas==1.4.2
pydantic
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
starlette==0.18.0
threadpoolctl==3.1.0
typing_extensions==4.2.0
uvicorn==0.17.6
spacy
pyenchant